ultimo-commit

+++ projects/migrations/0001_initial.py
@@ -18,7 +18,7 @@ class Migration(migrations.Migration):
                 ('nombre', models.CharField(max_length=200)),
                 ('descripcion', models.TextField()),
                 ('categoria', models.CharField(max_length=50)),
-                ('precio' , models.DecimalField(max_digits=10, decimal_places=2)),
+                ('precio' , models.IntegerField()),
                 ('stock' , models.IntegerField()),
                 ('created_at', models.DateTimeField(auto_now_add=True)),
             ],
