+++ projects/models.py
@@ -5,7 +5,7 @@ class Project(models.Model):
     nombre = models.CharField(max_length=200)
     descripcion = models.TextField()
     categoria = models.CharField(max_length=50)
-    precio = models.DecimalField(max_digits=10, decimal_places=2)
+    precio = models.IntegerField()
     stock = models.IntegerField()
     created_at =  models.DateTimeField(auto_now_add=True)
